+++ CHANGELOG.rst
@@ -3,6 +3,7 @@ Release history
 
 0.3.2 (November 2013)
 +++++++++++++++++++++
+- added: command line option --debug to increase loglevel 
 - fixed potential race issues in detector base class
 - fixed: several typos, test structure, naming conventions
 - more dynamic importing of detector code